initial commit
in this first commit, i prepare the environtment and tools to work on this small-fun todo list project. tools that i used is prettier as the formatter, eslint as the linter, and parcel as the bundler. i also added .gitignore to ignore some auto generate file/folders and node_modules